pyproject.toml
[project]
name = "adk-artifact"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.4.0",
    "matplotlib>=3.10.1",
]
